move configuration schema generation and validation
